eventserver: respect order of events while processing
Process events in order according to sequence numbers instead of the
order in which event source sends them. Order is preserved by
maintaining a map[int]Event and processing events according to global
variable currentEventSequence. currentEventSequence determines the
sequence number of event which we can present now. The server will wait
and keep storing all incoming events in the map until it receives the
event with sequence == currentEventSequence.